rustc_abi: move WrappingRange to its own file
